# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stories API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stories/{story_id}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получение списка комментариев к определенному сторису",
                "operationId": "ccd20c07f65807e93b3765fa5e29500f",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Связанные данные. Доступные связи: 'replies', 'emojis'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "replies"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка данных. Сортируемые поля: 'updated_at'. Если передать значение без знака, отсортирует в порядке возрастания, если со знаком минус (-updated_at), отсортирует в порядке убывания.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "-updated_at"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID комментария",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": ""
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "ID родительского комментария. В ответе будет список комментариев, которые являются ответами на комментарий с указанным parent_id.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "story_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "7000801"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Тестовый комментарий"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "16:48:54 27-09-2024"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer",
                                                        "example": null
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Создание комментария к определенному сторису",
                "operationId": "ec3bad3d10ef1a2574e867f5456af55d",
                "parameters": [
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "7000801"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Тестовый коммент"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "type": "string",
                                            "example": "CREATED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Комментарий к сторису был успешно создан"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "42"
                                                },
                                                "story_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "7000801"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "TEST TEST COMMENT"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "14:20:12 04-10-2024"
                                                },
                                                "parent_id": {
                                                    "type": "integer",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/stories/{story_id}/comments/{comment_id}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Получение комментария по ID к определенному сторису",
                "operationId": "735670e4ec17de69db1038ccd1cc188a",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 2
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Связанные данные. Доступные связи: 'replies', 'emojis'.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "replies"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "story_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "7000801"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Тестовый комментарий"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "16:48:54 27-09-2024"
                                                },
                                                "parent_id": {
                                                    "type": "integer",
                                                    "example": null
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Удаление комментария по ID у определенного сториса",
                "operationId": "e5a4f4e8678f9f14201476c720a68cd2",
                "parameters": [
                    {
                        "name": "story_id",
                        "in": "path",
                        "description": "ID сториса",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Номер телефона",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "7000801"
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "ID комментария",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "type": "string",
                                            "example": "DELETED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Комментарий к сторису был успешно удален"
                                        },
                                        "data": {
                                            "properties": {
                                                "phone": {
                                                    "type": "string",
                                                    "example": "7000801"
                                                },
                                                "story_id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "comment_id": {
                                                    "type": "integer",
                                                    "example": "42"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "name": "apiKey",
                "in": "header"
            }
        }
    },
...
}